Update CONTRIBUTORS.md (#54)
This PR updates the list of contributors with the persons that de facto
manages the github repository.

In the future, I think the list should also includes the person that
made significant contributions to the code.

Cheers,
Noé. 



<!-- Thank you so much for your contribution! -->
## PR Checklist

- [x] Pull request points to the `dev` branch.
- [x] If it is a bugfix, the PR should have label `on-merge:
backport-to-stable`.
- [x] For new code features, the PR should have been tested (please
include the results of the tests below or, when possible, add a new test
for the feature).