Cleanup (#83)
Move add_gravity_source_terms and add_pdv_source_terms out of set_uold,
which is closer to the logic behind the subroutine. They are now
directly in amr_step.
Updated this in the patches that were not too ancient.

Disable default writing particle memory stats for each cpu to prevent
spamming the log when nproc is high.

Cleanup pi.